Extract abstract base class for Editable based fields
